+++ link-move/src/test/java/com/nhl/link/move/valueconverter/LocalDateTimeConverterTest.java
@@ -16,14 +16,14 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class LocalDateTimeConverterTest {
 
-    private static final LocalDateTimeConverter normalizer = new LocalDateTimeConverter();
+    private static final LocalDateTimeConverter CONVERTER = new LocalDateTimeConverter();
 
     @Test
     public void testConvert_utilDate() {
         Instant now = Instant.now();
         LocalDateTime localDateTime = now.atZone(ZoneId.systemDefault()).toLocalDateTime().truncatedTo(ChronoUnit.MILLIS);
         Date date = new Date(now.toEpochMilli());
-        assertEquals(localDateTime, normalizer.convert(date));
+        assertEquals(localDateTime, CONVERTER.convert(date));
     }
 
     @Test
@@ -31,12 +31,12 @@ public class LocalDateTimeConverterTest {
         Instant now = Instant.now();
         LocalDateTime localDateTime = now.atZone(ZoneId.systemDefault()).toLocalDateTime().truncatedTo(ChronoUnit.MILLIS);
         java.sql.Date date = new java.sql.Date(now.toEpochMilli());
-        assertEquals(localDateTime, normalizer.convert(date));
+        assertEquals(localDateTime, CONVERTER.convert(date));
     }
 
     @Test
     public void testConvert_sqlTime() {
-        assertThrows(LmRuntimeException.class, () -> normalizer.convert(new Time(Instant.now().toEpochMilli())));
+        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(new Time(Instant.now().toEpochMilli())));
     }
 
     @Test
@@ -44,6 +44,11 @@ public class LocalDateTimeConverterTest {
         Instant now = Instant.now();
         LocalDateTime localDateTime = now.atZone(ZoneId.systemDefault()).toLocalDateTime().truncatedTo(ChronoUnit.MILLIS);
         Timestamp timestamp = new Timestamp(now.toEpochMilli());
-        assertEquals(localDateTime, normalizer.convert(timestamp));
+        assertEquals(localDateTime, CONVERTER.convert(timestamp));
+    }
+
+    @Test
+    public void testConvert_string() {
+        assertEquals(LocalDateTime.of(2017, 1, 2, 1, 0, 1), CONVERTER.convert("2017-01-02T01:00:01"));
     }
 }